pyro_camera_api/pyro_camera_api/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==4.11.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.8 ; python_version == "3.9"
click==8.3.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
coloredlogs==15.0.1 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.3.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.110.3 ; python_version >= "3.9" and python_version < "4.0"
flatbuffers==25.9.23 ; python_version >= "3.9" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.7.1 ; python_version >= "3.9" and python_version < "4.0"
humanfriendly==10.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.11 ; python_version >= "3.9" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
ncnn==1.0.20240410 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.9" and python_version < "4.0"
onnxruntime==1.19.2 ; python_version >= "3.9" and python_version < "4.0"
opencv-python-headless==4.11.0.86 ; python_version >= "3.9" and python_version < "4.0"
opencv-python==4.11.0.86 ; python_version >= "3.9" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.9" and python_version < "4.0"
portalocker==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==6.33.1 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.41.5 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.12.4 ; python_version >= "3.9" and python_version < "4.0"
pyreadline3==3.5.4 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
python-dotenv==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
pywin32==311 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
pyyaml==6.0.3 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.9" and python_version < "4.0"
sympy==1.14.0 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.9" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.30.6 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.22.1 ; python_version >= "3.9" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.9" and python_version < "4.0"
